refactor: refactor files pathes
