+++ package.json
@@ -13,10 +13,12 @@
     "test": "jest"
   },
   "dependencies": {
+    "@reduxjs/toolkit": "^1.9.7",
     "jest-environment-jsdom": "^29.7.0",
     "node-fetch": "^2.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-redux": "^8.1.3",
     "react-router-dom": "^6.18.0",
     "ts-node": "^10.9.1"
   },
